Add alternate method of generating guard spawn times
- Spawns them linearly instead of in large chunks

+++ src/Xyz/Assets/Xyz/Scripts/Singleton/DifficultyEvent.cs
@@ -7,7 +7,7 @@
         AddPushers
     }
 
-    public int Time { get; private set; }
+    public float Time { get; private set; }
     public string Message { get; private set; }
 
     public Type EventType { get; private set; }
@@ -15,7 +15,7 @@
     public bool Fired { get; set; }
     
 
-    public DifficultyEvent(int time, string message, Type eventType)
+    public DifficultyEvent(float time, string message, Type eventType)
     {
         Time = time;
         Message = message;
